update cardDetails,navbar,mycart route

+++ src/Component/MyCart/MyCart.jsx
@@ -9,34 +9,34 @@ const MyCart = () => {
   const [productCart, setProductCart] = useState(carts);
   console.log(user.email);
   console.log(productCart);
-  const filterCart = productCart.filter(item => item.user.email == user.email);
+  const filterCart = productCart.filter(item => item.email == user.email);
   console.log(filterCart);
   const handleDelete = _id => {
     console.log(_id);
-    // Swal.fire({
-    //   title: 'Are you sure?',
-    //   text: "You won't be able to revert this!",
-    //   icon: 'warning',
-    //   showCancelButton: true,
-    //   confirmButtonColor: '#3085d6',
-    //   cancelButtonColor: '#d33',
-    //   confirmButtonText: 'Yes, delete it!',
-    // }).then(result => {
-    // if (result.isConfirmed) {
-    fetch(`http://localhost:5000/deleteCard/${_id}`, {
-      method: 'DELETE',
-    })
-      .then(res => res.json())
-      .then(data => {
-        console.log(data);
-        if (data.deletedCount > 0) {
-          Swal.fire('Deleted!', 'Your file has been deleted.', 'success');
-          const remaining = filterCart.filter(item => item._id == _id);
-          setProductCart(remaining);
-        }
-      });
-    // }
-    // });
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+    }).then(result => {
+      if (result.isConfirmed) {
+        fetch(`http://localhost:5000/deleteCard/${_id}`, {
+          method: 'DELETE',
+        })
+          .then(res => res.json())
+          .then(data => {
+            console.log(data);
+            if (data.deletedCount > 0) {
+              Swal.fire('Deleted!', 'Your file has been deleted.', 'success');
+              const remaining = filterCart.filter(item => item._id !== _id);
+              setProductCart(remaining);
+            }
+          });
+      }
+    });
   };
   return (
     <div className="w-10/12 mx-auto">
@@ -46,7 +46,7 @@ const MyCart = () => {
           <div className=" px-5 py-5 bg-orange-300 rounded-lg" key={cart._id}>
             <img className="w-full h-56 rounded-lg" src={cart.image} alt="" />
             <h3>{cart.name}</h3>
-            <h3>{cart.user.email}</h3>
+            <h3>{cart.email}</h3>
             <button
               onClick={() => handleDelete(cart._id)}
               className="btn btn-error"
